Used Copilot to assist me with coverting a date string format within XAML. I did vibe code it by prompting what I needed.
I fully understand how the c# is doing it, just not 100% sure how XAML is retrieving the method. It appears the interface is figuring it out based on the arguments used. Need to research interfaces more. Going to convert my database methods into an interface when my calculator is finished.